requirements.txt
asgiref==3.2.7
boto3==1.16.5
botocore==1.19.5
Django==3.0.5
django-crispy-forms==1.11.1
django-developer-panel==0.1.2
django-filter==2.3.0
django-storages==1.10.1
gunicorn==20.0.4
jmespath==0.10.0
Pillow==7.2.0
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.11
whitenoise==5.2.0
